front: rename spaceTimeChart folder to SpaceTimeChartWrapper
