+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/ManchetteMenuButton.tsx
