+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/ProjectionLoadingMessage.tsx
