+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/SettingsPanel.tsx
