+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/WaypointsPanel.tsx
