+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/helpers/__tests__/makeProjectedItems.spec.ts
