+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/helpers/getPathStyle.ts
