+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/helpers/makeProjectedItems.ts
