+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/helpers/utils.ts
