+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/helpers/zones.ts
