+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/useProjectedConflicts.ts
