+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/useTrackOccupancy.ts
