+++ front/src/modules/simulationResult/components/SpaceTimeChartWrapper/useWaypointMenu.tsx
